+++ code/test_2D_fully.py
@@ -53,7 +53,7 @@ def test_single_volume(case, net, test_save_path, FLAGS):
             0).unsqueeze(0).float()
         
         if torch.backends.mps.is_available():
-            input = input.to("mps")
+            input = input.to(torch.float32).to("mps")
         else:
             input = input.cuda()
